Add files via upload
This is a program that generates a random integer between 0 and 100. It then requests for user input as a guess to what the number could be. If it is not the number generated then the output will say that the guess is either to high (larger than the number generated) or to low. The program then asks the user again to guess the number, repeating the process until the user enters the correct integer, thus terminating the program